feat: デッキ管理機能を改善するためのユーティリティ関数を追加し、重複チェックとデッキデータ生成を統合
